backend/common/pyproject.toml
[project]
name = "ai-marketing-common"
version = "0.1.0"
description = "Shared modules for AI Marketing Asset Engine microservices"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "asyncpg>=0.29.0",
  "alembic>=1.13.2",
  "python-jose[cryptography]>=3.3.0",
  "google-auth>=2.34.0",
  "redis>=5.0.8",
  "httpx>=0.27.0",
  "structlog>=24.4.0",
  "requests>=2.32.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["common*"]
